move ranking registration transform
