+++ src/app/ranking/pages/RankingProfile.tsx
@@ -9,7 +9,7 @@ import ContestApi from '../../contest/api'
 import ContestLogsByDayGraph from '../components/graphs/ReadingActivityGraph'
 import ContestLogsByMediumGraph from '../components/graphs/MediaDistributionGraph'
 import ContestLogsOverview from '../components/ContestLogsOverview'
-import { rankingsToRegistrationOverview } from '../transform/graph'
+import { rankingsToRegistrationOverview } from '../transform/ranking'
 import { Contest } from '../../contest/interfaces'
 import { useCachedApiState, isReady } from '../../cache'
 import { contestSerializer } from '../../contest/transform'
